inlong-sort/requirements.txt
kafka-python==2.0.2
elasticsearch==7.14.0
requests==2.31.0
